Proposed fix for aAsserting for interfaces resulting in Ambiguous method error under Java 8 #365
